A couple of tweaks
