+++ src/Watson/Sitemap/Sitemap.php
@@ -111,13 +111,14 @@ class Sitemap
      * @param  \DateTime|string  $lastModified
      * @param  string  $changeFrequency
      * @param  string  $priority
-     * @return Tag
+     * @return \Watson\Sitemap\Tags\Tag
      */
     public function addTag($location, $lastModified = null, $changeFrequency = null, $priority = null)
     {
         $tag = $location instanceof Tag ? $location : new Tag($location, $lastModified, $changeFrequency, $priority);
 
         $this->tags[] = $tag;
+
         return $tag;
     }
 
@@ -176,15 +177,7 @@ class Sitemap
             return response()->make($cachedView, 200, ['Content-type' => 'text/xml']);
         }
 
-        $hasImages = false;
-        foreach ($this->tags as $tag) {
-            if ($tag->hasImages()) {
-                $this->hasImages = true;
-                break;
-            }
-        }
-
-        $sitemap = response()->view('sitemap::sitemap', ['__tags' => $this->getTags(), '__hasImages' => $this->hasImages], 200, ['Content-type' => 'text/xml']);
+        $sitemap = response()->view('sitemap::sitemap', ['__tags' => $this->getTags(), '__hasImages' => $this->imagesPresent()], 200, ['Content-type' => 'text/xml']);
 
         $this->saveCachedView($sitemap);
 
@@ -242,6 +235,22 @@ class Sitemap
             $key = $this->getCacheKey();
 
             return $this->cache->has($key);
+        }
+
+        return false;
+    }
+
+    /**
+     * Return whether there are any images present in the sitemap.
+     *
+     * @return bool
+     */
+    protected function imagesPresent()
+    {
+        foreach ($this->tags as $tag) {
+            if ($tag->hasImages()) {
+                return true;
+            }
         }
 
         return false;
