+++ src/views/sitemaps.php
@@ -5,7 +5,7 @@
     <loc><?php echo htmlspecialchars($__sitemap->getLocation(), ENT_XML1) ?></loc>
     <?php if ($__sitemap->getLastModified()): ?>
     <lastmod><?php echo $__sitemap->getLastModified()->format('Y-m-d\TH:i:sP') ?></lastmod>
-    <?php endif; ?>
+    <?php endif ?>
   </sitemap>
   <?php endforeach ?>
 </sitemapindex>
